+++ src/Commands/DomainDtoMakeCommand.php
@@ -2,8 +2,12 @@
 
 namespace Lunarstorm\LaravelDDD\Commands;
 
+use Lunarstorm\LaravelDDD\Commands\Concerns\HasDomainStubs;
+
 class DomainDtoMakeCommand extends DomainGeneratorCommand
 {
+    use HasDomainStubs;
+
     protected $name = 'ddd:dto';
 
     /**
@@ -28,7 +32,7 @@ class DomainDtoMakeCommand extends DomainGeneratorCommand
 
     protected function getStub()
     {
-        return $this->resolveStubPath('dto.php.stub');
+        return $this->resolveDddStubPath('dto.stub');
     }
 
     protected function getRelativeDomainNamespace(): string