+++ src/Commands/DomainMailMakeCommand.php
@@ -3,11 +3,13 @@
 namespace Lunarstorm\LaravelDDD\Commands;
 
 use Illuminate\Foundation\Console\MailMakeCommand;
+use Lunarstorm\LaravelDDD\Commands\Concerns\HasDomainStubs;
 use Lunarstorm\LaravelDDD\Commands\Concerns\ResolvesDomainFromInput;
 
 class DomainMailMakeCommand extends MailMakeCommand
 {
-    use ResolvesDomainFromInput;
+    use HasDomainStubs,
+        ResolvesDomainFromInput;
 
     protected $name = 'ddd:mail';
 }