+++ src/Facades/DDD.php
@@ -3,12 +3,15 @@
 namespace Lunarstorm\LaravelDDD\Facades;
 
 use Illuminate\Support\Facades\Facade;
+use Lunarstorm\LaravelDDD\StubManager;
 
 /**
  * @see \Lunarstorm\LaravelDDD\DomainManager
  *
  * @method static void filterAutoloadPathsUsing(callable $filter)
  * @method static void resolveNamespaceUsing(callable $resolver)
+ * @method static string packagePath(string $path = '')
+ * @method static StubManager stubs()
  */
 class DDD extends Facade
 {