+++ src/LaravelDDDServiceProvider.php
@@ -27,6 +27,8 @@ class LaravelDDDServiceProvider extends PackageServiceProvider
             ->hasConfigFile()
             ->hasCommands([
                 Commands\InstallCommand::class,
+                Commands\PublishCommand::class,
+                Commands\StubCommand::class,
                 Commands\UpgradeCommand::class,
                 Commands\OptimizeCommand::class,
                 Commands\OptimizeClearCommand::class,
@@ -61,12 +63,37 @@ class LaravelDDDServiceProvider extends PackageServiceProvider
                 Commands\Migration\DomainMigrateMakeCommand::class,
             ]);
 
-        if (app()->version() >= 11) {
+        if ($this->laravelVersion(11)) {
             $package->hasCommand(Commands\DomainClassMakeCommand::class);
             $package->hasCommand(Commands\DomainEnumMakeCommand::class);
             $package->hasCommand(Commands\DomainInterfaceMakeCommand::class);
             $package->hasCommand(Commands\DomainTraitMakeCommand::class);
         }
+
+        // if ($this->laravelVersion('11.30.0')) {
+        //     $package->hasCommand(Commands\PublishCommand::class);
+        //     $package->hasCommand(Commands\StubCommand::class);
+        // }
+    }
+
+    protected function laravelVersion($value)
+    {
+        return version_compare(app()->version(), $value, '>=');
+    }
+
+    protected function registerMigrations()
+    {
+        $this->app->singleton(Commands\Migration\DomainMigrateMakeCommand::class, function ($app) {
+            // Once we have the migration creator registered, we will create the command
+            // and inject the creator. The creator is responsible for the actual file
+            // creation of the migrations, and may be extended by these developers.
+            $creator = $app['migration.creator'];
+            $composer = $app['composer'];
+
+            return new Commands\Migration\DomainMigrateMakeCommand($creator, $composer);
+        });
+
+        $this->loadMigrationsFrom(DomainMigration::paths());
     }
 
     protected function registerMigrations()
@@ -87,14 +114,14 @@ class LaravelDDDServiceProvider extends PackageServiceProvider
     public function packageBooted()
     {
         $this->publishes([
-            $this->package->basePath('/../stubs') => resource_path("stubs/{$this->package->shortName()}"),
+            $this->package->basePath('/../stubs') => $this->app->basePath("stubs/{$this->package->shortName()}"),
         ], "{$this->package->shortName()}-stubs");
 
         if ($this->app->runningInConsole() && method_exists($this, 'optimizes')) {
             $this->optimizes(
                 optimize: 'ddd:optimize',
                 clear: 'ddd:clear',
-                key: 'ddd cache',
+                key: 'laravel-ddd',
             );
         }
     }