+++ tests/Factory/DomainFactoryTest.php
@@ -1,9 +1,11 @@
 <?php
 
 use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\Config;
 use Lunarstorm\LaravelDDD\Factories\DomainFactory;
+use Lunarstorm\LaravelDDD\Support\DomainAutoloader;
 
 it('can resolve the factory name of a domain model', function ($modelClass, $expectedFactoryClass) {
     $this->setupTestApplication();
@@ -26,8 +28,15 @@ it('is backwards compatible with factories located in database/factories/**/*', 
 ]);
 
 it('can instantiate a domain model factory', function ($domainParameter, $modelName, $modelClass) {
+    $this->afterApplicationCreated(function () {
+        (new DomainAutoloader)->autoload();
+    });
+
+    $this->setupTestApplication();
+
     Config::set('ddd.base_model', 'Lunarstorm\LaravelDDD\Models\DomainModel');
     Artisan::call("ddd:model -f {$domainParameter}:{$modelName}");
+
     expect(class_exists($modelClass))->toBeTrue();
     expect($modelClass::factory())->toBeInstanceOf(Factory::class);
 })->with([